0x05. C - Pointers, arrays and strings

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -1,4 +1,5 @@
 #include "main.h"
+
 /**
  * _atoi - convert a string to an integer.
  * @s: pointer to string
@@ -7,17 +8,20 @@
  */
 int _atoi(char *s)
 {
-int sign = 1;
-int result = 0;
-while (*s)
-{
-if (*s == '-')
-sign = -sign;
-else if (*s >= '0' && *s <= '9')
-result = result * 10 + (*s - '0');
-else if (result > 0)
-break;
-s++;
-}
-return (sign *result);
+    int sign = 1;
+    unsigned int result = 0;
+
+    while (*s)
+    {
+        if (*s == '-')
+            sign = -sign;
+        else if (*s >= '0' && *s <= '9')
+            result = result * 10 + (*s - '0');
+        else if (result > 0)
+            break;
+
+        s++;
+    }
+
+    return sign * (int)result;
 }
